requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
tzdata>=2024.1